+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## master
 
+- react-{identicon, qr}, ui-{keyring, settings}, vue-identicon 3.2.1
 - phishing 0.21.1
 - hw-ledger, keyring, util, util-crypto 12.0.1
 - wasm-crypto 7.1.1